Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -7,6 +7,6 @@ char* welcome() {
 int main() {
     int a;
     scanf("%d",&a);
-    printf("Area: %.2f", 3.14*float(a*a));
+    printf("Area: %.2f", 3.14*(a*a));
     return 0;
 }
